# refactor: faster