--- a/relayer/rly/cosmos_relayer.go
+++ b/relayer/rly/cosmos_relayer.go
@@ -6,6 +6,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"strings"
+	"time"
 
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	"github.com/docker/docker/client"
@@ -42,18 +43,19 @@ func NewCosmosRelayer(log *zap.Logger, testName string, cli *client.Client, netw
 }
 
 type CosmosRelayerChainConfigValue struct {
-	AccountPrefix  string  `json:"account-prefix"`
-	ChainID        string  `json:"chain-id"`
-	Debug          bool    `json:"debug"`
-	GRPCAddr       string  `json:"grpc-addr"`
-	GasAdjustment  float64 `json:"gas-adjustment"`
-	GasPrices      string  `json:"gas-prices"`
-	Key            string  `json:"key"`
-	KeyringBackend string  `json:"keyring-backend"`
-	OutputFormat   string  `json:"output-format"`
-	RPCAddr        string  `json:"rpc-addr"`
-	SignMode       string  `json:"sign-mode"`
-	Timeout        string  `json:"timeout"`
+	AccountPrefix   string        `json:"account-prefix"`
+	ChainID         string        `json:"chain-id"`
+	Debug           bool          `json:"debug"`
+	GRPCAddr        string        `json:"grpc-addr"`
+	GasAdjustment   float64       `json:"gas-adjustment"`
+	GasPrices       string        `json:"gas-prices"`
+	Key             string        `json:"key"`
+	KeyringBackend  string        `json:"keyring-backend"`
+	OutputFormat    string        `json:"output-format"`
+	RPCAddr         string        `json:"rpc-addr"`
+	SignMode        string        `json:"sign-mode"`
+	Timeout         string        `json:"timeout"`
+	MinLoopDuration time.Duration `json:"min-loop-duration" default:"100ms"`
 }
 
 type CosmosRelayerChainConfig struct {
@@ -83,18 +85,19 @@ func ChainConfigToCosmosRelayerChainConfig(chainConfig ibc.ChainConfig, keyName,
 	return CosmosRelayerChainConfig{
 		Type: chainType,
 		Value: CosmosRelayerChainConfigValue{
-			Key:            keyName,
-			ChainID:        chainConfig.ChainID,
-			RPCAddr:        rpcAddr,
-			GRPCAddr:       gprcAddr,
-			AccountPrefix:  chainConfig.Bech32Prefix,
-			KeyringBackend: keyring.BackendTest,
-			GasAdjustment:  chainConfig.GasAdjustment,
-			GasPrices:      chainConfig.GasPrices,
-			Debug:          true,
-			Timeout:        "10s",
-			OutputFormat:   "json",
-			SignMode:       "direct",
+			Key:             keyName,
+			ChainID:         chainConfig.ChainID,
+			RPCAddr:         rpcAddr,
+			GRPCAddr:        gprcAddr,
+			AccountPrefix:   chainConfig.Bech32Prefix,
+			KeyringBackend:  keyring.BackendTest,
+			GasAdjustment:   chainConfig.GasAdjustment,
+			GasPrices:       chainConfig.GasPrices,
+			Debug:           true,
+			Timeout:         "10s",
+			OutputFormat:    "json",
+			SignMode:        "direct",
+			MinLoopDuration: 100 * time.Millisecond,
 		},
 	}
 }