## Summary

Make Interchaintest Fast like zooooooooooooooomies

## Benchmarks
```
// TODO: time to beat:
 // Before this PR: 103.31s
 // With faster blocks (in this PR) & faster relayer polling: 70.39s / 67.03s 
// Without startup Locks: 62.8s
```