@Reecepbcups 2024-08-07
Build (including relayer setup) takes 50s of the total 60s test time. If we can speed up the ic.Build we could get test down to an absolute best case of ~15s (down from 103s)