@Reecepbcups 2024-08-08
https://github.com/strangelove-ventures/interchaintest/issues/1153 is probably the next big performance improvement. Getting relayers to auto provision for planned testing is ideal, lot of dead time waiting.

We already know the chain configs to `chains add`, `rly keys restore`, and `rly paths new` (config). Then we just `rly tx link` as soon as the chain is at height capable.

Must allow relayers pre ic.Build() standalone