@Reecepbcups 2024-08-08
ideas from Andrew:
- sleep inifinite then just exec into the container on start