@Reecepbcups 2024-08-13
export genesis state after IBC start, then can load back from that in future runs